feat: 프로그램 세부 정보 팝업 추가
